requirements.txt
django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
unidecode==1.1.1
